add latches to touch
The issue latches solve is it's possible we get touchStart, touchMove,
touchEnd, all within the same frame so from the games POV it would see
nothing because the user let up before we had a chance to see it.
The latches keeps the value around until we read it. A complication is
if we see opposite directions we have to remove the oldest reading.
In other words if we see R then L we need to erase R from the latch.