Fix Keyboard Input
There's got to be a simpler way and maybe it should be done
at a higher level. The difference between a keyboard and
and joypad is that on a joypad it's impossible to press
opposite directions at the same time.

On keyboard if we press WITHOUT RELEASING Up then Right Then Down
the play pressed Down last so they want to go down. If they then
release Down we know they are still pressing Up and Right so which
way do they want to go? Up or Right? Since Right was pressed after
Up they arguably want to go Right so we need some way to remember
Right is a newer input than up.

Further, we have the game loop. If a player is able to press and release
up faster than the game loop the game, looking at the state of the
keys, would see no keys being pressed. That means we need to save
the fact that a key was presssed.

I think my gut would say that at the lowest level, keyboard code should
should record which keys were pressed and which keys are currently down.
(2 bits per key) but 2 bits per key are not enough to handle the
cases above. So, this code is complicated and crap. Further, It has to
treat directions special from other keys (fire).

I think I'm not seeing the simpler solution 😕