Add bombs
I'm not sure if enemies should explode bombs or not.

One the one hand, bombs seem inert. On the other hand
it feels like they wouldn't be all that useful if the
bombs didn't explode when enemies touch them because
if they don't then what's the difference between a
bomb and a rock. If the bombs don't explode when an
enemy touches them then you can mostly only drop them
on enemies which is the same as rocks. You could maybe
use them to open holes but that feels like a limited
use.